+++ Source/SwiftLintBuiltInRules/Rules/RuleConfigurations/ImplicitOptionalInitConfiguration.swift
@@ -1,0 +1,17 @@
+import SwiftLintCore
+
+@AutoConfigParser
+struct ImplicitOptionalInitConfiguration: SeverityBasedRuleConfiguration {
+  typealias Parent = ImplicitOptionalInitializationRule
+
+  @AcceptableByConfigurationElement
+  enum Style: String {
+    case always
+    case never
+  }
+
+  @ConfigurationElement(key: "severity")
+  private(set) var severityConfiguration = SeverityConfiguration<Parent>(.warning)
+  @ConfigurationElement(key: "style")
+  private(set) var style: Style = .always
+}